requirements.txt
pytest-playwright
PyYAML
typer